Create pools.py
