+++ booru/views/__init__.py
@@ -4,3 +4,4 @@ from .misc import *
 from .posts import *
 from .tags import *
 from .users import *
+from .pools import *
